requirements.txt
agenti
requests
pyyaml